Move a big chunk of the jitdump code to wasmtime-jit-debug
