+++ crates/jit-debug/Cargo.toml
@@ -12,9 +12,12 @@ edition = "2018"
 
 [dependencies]
 lazy_static = {version = "1.3.0", optional = true }
+libc = { version = "0.2.60", default-features = false, optional = true }
+scroll = { version = "0.10.1", features = ["derive"], optional = true }
 
 [badges]
 maintenance = { status = "actively-developed" }
 
 [features]
 gdb_jit_int = ["lazy_static"]
+perf_jitdump = ["scroll", "libc"]
