+++ crates/jit-debug/src/lib.rs
@@ -1,2 +1,5 @@
 #[cfg(feature = "gdb_jit_int")]
 pub mod gdb_jit_int;
+
+#[cfg(all(feature = "perf_jitdump", target_os = "linux"))]
+pub mod perf_jitdump;
