+++ crates/profiling/Cargo.toml
@@ -16,11 +16,10 @@ cfg-if = "1.0"
 gimli = { version = "0.23.0", optional = true }
 lazy_static = "1.4"
 libc = { version = "0.2.60", default-features = false }
-scroll = { version = "0.10.1", features = ["derive"], optional = true }
-serde = { version = "1.0.99", features = ["derive"] }
 target-lexicon = "0.11.0"
 wasmtime-environ = { path = "../environ", version = "0.25.0" }
 wasmtime-runtime = { path = "../runtime", version = "0.25.0" }
+wasmtime-jit-debug = { path = "../jit-debug", version = "0.25.0", features = ["perf_jitdump"] }
 ittapi-rs = { version = "0.1.5", optional = true }
 
 [dependencies.object]
@@ -33,5 +32,5 @@ features = ['read_core', 'elf', 'std']
 maintenance = { status = "actively-developed" }
 
 [features]
-jitdump = ['object', 'scroll', 'gimli']
+jitdump = ['object', 'gimli']
 vtune = ['ittapi-rs']
